Random_python_tests/requirements.txt
opencv-python
mediapipe
numpy
Pillow
transformers
torch
pyautogui
openai<1.0
uvicorn
fastapi
pydantic
sse-starlette
transformers
einops
tiktoken
transformers_stream_generator
accelerate>=0.26.0
transformers
scikit-learn
nltk
spacy
pyyaml
requests
